Retrieve table columns as array of Column objects (revised) _This version does not require any changes to existing tests._
When using the Autoset trait, the getDbColumnTypes() method uses many separate queries to get the column type and column default.

This change retrieves the table columns just once as an array of \Doctrine\DBAL\Schema\Column and iterates through that array to retrieve the column type and default. This results in a fairly dramatic performance improvement.