Fix image field when form is re-displayed.
When trying to create a new record and validation fails, the form is displayed
again with the submitted values for the user to fix the issues. The images
however are handled a bit differently: they're displayed inline as
`data:image/sometype;...` so that the user doesn't need to select them again
from disk.

The issue here is that the base URL gets added in front of
`data:image/sometype`, which causes the display of the image to fail since
`http://somedomain/data:image/sometype;...` isn't valid.