Updated YUMMY License to be more clear, prevent emails and mention punishments
Not much is different from what we've done so far, but we're now formalising it into the license text so that hopefully people read it instead of asking us about it, and that we have a text to reference for every question we've had in v3. We've added article 14 and 15 - punishment when found to be using Backpack illegally. That one is new.