addresses #2266 - fixed preview in image and base64_image fields
When you included multiple fields of this type, using Javascript instead of PHP (for ex inside the repeatable field type), the cropper always showed the preview of the first element, instead of each instance of the field having its own preview area. This fixes that by making the selector use classes instead of IDs.