fix #3108 - move Show operation setup to setupShowOperation method
In Backpack 4.0 and 4.1, in order to keep backwards-compatibility, we've kept a behaviour for the Show operation that is not intuitive at all. The operation tries to guess which columns to show, if none were provided (which is great), but because it does that, it makes it difficult for developer to remove columns from the Show operation. This commit fixes it, by moving that logic inside setupShowOperation, which is automatically called. It also opens up the opportunity for developers to completely overwrite this behaviour (by creating a new method with this name - setupShowOperation) in their controllers, or to insert code right in the middle of this logic (by creating a method called modifyShowOperation). More details inside https://github.com/Laravel-Backpack/CRUD/issues/3108