default user model is now the one in auth
Because in Laravel 8+ the default User model has been moved from App\User to App\Models\User, by default when you install Backpack 4.1.x on Laravel 8, you won't be able to login. This is a very bad first experience when installing Backpack, so we should make it "just work" both on Laravel 6 and 7 (App\User) and on Laravel 8 (App\Models\User). One way would be to just use the same config value in the auth.php file - which is what this change does.